moves plugin types to separat file
