+++ packages/types/src/chain.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from "ethers";
 
-import { DeployedPlugins } from "./config";
+import { DeployedPlugins } from "./plugin";
 import { IrmTypes, OracleTypes, SupportedChains } from "./enums";
 import { FundingStrategy, LiquidationDefaults, RedemptionStrategy } from "./liquidation";
