+++ packages/types/src/index.ts
@@ -1,7 +1,8 @@
-export * from "./enums";
-export * from "./utils";
-export * from "./liquidation";
-export * from "./chain";
-export * from "./fuse";
-export * from "./config";
 export * from "./artifact";
+export * from "./chain";
+export * from "./config";
+export * from "./enums";
+export * from "./fuse";
+export * from "./liquidation";
+export * from "./plugin";
+export * from "./utils";
